+++ examples/configs/src/index.js
@@ -87,7 +87,7 @@ import { threeMinimalLight } from './view-configs/3d-maps/three-minimal-light.js
 import { linkControllerDemo } from './view-configs/3d-maps/link-controller.js';
 import { linkControllerMinimal } from './view-configs/3d-maps/link-controller-minimal.js';
 import { neuroglancerTwoLayers } from './view-configs/3d-maps/neuroglancer-two-layers.js';
-import { neuroglancerMocosko } from './view-configs/3d-maps/neuroglancer-mocosco.js'
+import { neuroglancerMocosko } from './view-configs/3d-maps/neuroglancer-mocosko.js';
 // Nature Methods figures
 import { citeSeq } from './view-configs/paper-figures/cite-seq.js';
 import { codex } from './view-configs/paper-figures/codex.js';
